refactor(staking): reorganize rewards chart files into own subfolder
