+++ packages/staking/src/features/activity/PastEpochsRewards/PoolIndicator.tsx
