+++ packages/staking/src/features/activity/PastEpochsRewards/index.ts
@@ -1,0 +1,1 @@
+export { PastEpochsRewards } from './PastEpochsRewards';
